refactor(core): move mqtt_data to use_cases data
